Partner Service delete done.

+++ app/Http/Controllers/JobServiceController.php
@@ -177,8 +177,11 @@ class JobServiceController extends Controller
     {
         try {
             $job_service = JobService::find($job_service);
+            if (!$job_service) return api_response($request, null, 400, ['message' => "Job service not found"]);
             $response = (new JobServiceActions)->delete($job_service);
-            if ($response['code'] == 400) return api_response($response, null, 421, ['message' => $response['msg']]);
+            if ($response['code'] == 400) return api_response($request, null, 400, ['message' => $response['msg']]);
+            elseif ($response['code'] == 200) return api_response($request, null, 200);
+            else return api_response($request, null, 500);
         } catch (\Throwable $exception) {
             app('sentry')->captureException($exception);
             return api_response($request, null, 500);
